fix: set a fixed height
