+++ src/webapp/pages/app/App.css
@@ -3,7 +3,6 @@ html {
     font-family: Roboto, Arial, sans-serif;
     overflow-y: scroll;
     font-weight: 300;
-    height: 100%;
 }
 
 body {
@@ -14,11 +13,4 @@ body {
 
 li {
     line-height: 1.75;
-}
-
-#root {
-    height: 100%;
-}
-#app {
-    height: 100%;
 }
